requirements.txt
kafka-python==2.0.2
six>=1.16.0
python-dateutil>=2.8.2